#37939 -- move to double math

+++ mvp/client/prover_mvp_android/app/src/main/cpp/BoundsChecker.cpp
@@ -43,8 +43,8 @@ bool BoundsChecker::CheckBounds(Vector current) {
     // turn _current so we should move to +x (of +x, +y for Diagonal)
     // fail if we've got too close to another swipe-point
     current.Mul(_turnMat);
-    float x = current._x;
-    float y = current._y;
+    double x = current._x;
+    double y = current._y;
 
     if (_isDiagonal) {
         if (x < -FIT_FACTOR_H || y < -FIT_FACTOR_H)
@@ -59,12 +59,12 @@ bool BoundsChecker::CheckBounds(Vector current) {
             current._x = x;
         }
         //distance to line x = y -- it is the line that goes to target point
-        float d1 = (x - y) / _sqrt2;
+        double d1 = (x - y) / _sqrt2;
         // distance to remaining non-target swipe point
-        float r2 = current.DistanceTo(1, 0);
+        double r2 = current.DistanceTo(1, 0);
 
         return d1 / FIT_FACTOR_D < r2 / (1 - FIT_FACTOR_D);
     } else {
-        return x >= -FIT_FACTOR_H && fabsf(y) <= FIT_FACTOR_H;
+        return x >= -FIT_FACTOR_H && fabs(y) <= FIT_FACTOR_H;
     }
 }
